Extract print_fixes
