+++ qlty-cli/src/ui/text.rs
@@ -1,20 +1,17 @@
-use anyhow::{Context as _, Result};
-use console::{style, Style};
-use dialoguer::theme::ColorfulTheme;
-use dialoguer::Input;
-use diffy::Patch;
+use anyhow::Result;
+use console::style;
 use num_format::{Locale, ToFormattedString as _};
 use qlty_analysis::utils::fs::path_to_string;
 use qlty_check::Report;
 use qlty_check::{executor::InvocationStatus, results::FixedResult};
 use qlty_config::Workspace;
-use qlty_types::analysis::v1::{ExecutionVerb, Issue, Level, SuggestionSource};
-use similar::{ChangeTag, TextDiff};
-use std::fmt;
-use std::io::{IsTerminal as _, Write};
+use qlty_types::analysis::v1::{ExecutionVerb, Issue};
+use std::io::Write;
 use tabwriter::TabWriter;
-use tracing::warn;
 
+use super::fixes::print_fixes;
+use super::level::formatted_level;
+use super::source::formatted_source;
 use super::unformatted::print_unformatted;
 
 #[derive(Debug)]
@@ -44,22 +41,16 @@ impl TextFormatter {
     }
 }
 
-struct Line(Option<usize>);
-
-impl fmt::Display for Line {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self.0 {
-            None => write!(f, "    "),
-            Some(idx) => write!(f, "{:<4}", idx + 1),
-        }
-    }
-}
-
 impl TextFormatter {
     pub fn write_to(&mut self, writer: &mut dyn std::io::Write) -> anyhow::Result<()> {
         if !self.summary {
             print_unformatted(writer, &self.report.issues)?;
-            self.print_fixes(writer)?;
+            print_fixes(
+                writer,
+                &self.report.issues,
+                &self.workspace.root,
+                self.apply_mode,
+            )?;
             self.print_issues(writer)?;
         }
 
@@ -69,16 +60,7 @@ impl TextFormatter {
     }
 }
 
-struct PatchCandidate {
-    issue: Issue,
-    source: SuggestionSource,
-    path: String,
-    patch: String,
-    original_code: String,
-    modified_code: String,
-}
-
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Copy)]
 pub enum ApplyMode {
     All,
     None,
@@ -265,170 +247,6 @@ impl TextFormatter {
         Ok(())
     }
 
-    pub fn print_fixes(&mut self, writer: &mut dyn std::io::Write) -> Result<()> {
-        let mut patch_candidates = vec![];
-
-        for issue in &self.report.issues {
-            if let Some(location) = &issue.location {
-                if let Some(suggestion) = issue.suggestions.first() {
-                    if let Ok(patch) = Patch::from_str(&suggestion.patch) {
-                        let full_path = self.workspace.root.join(location.path.clone());
-                        let original_code =
-                            std::fs::read_to_string(&full_path).with_context(|| {
-                                format!("Failed to read file: {}", full_path.display())
-                            })?;
-
-                        if let Ok(modified_code) = diffy::apply(&original_code, &patch) {
-                            patch_candidates.push(PatchCandidate {
-                                issue: issue.clone(),
-                                source: SuggestionSource::try_from(suggestion.source)
-                                    .unwrap_or_default(),
-                                path: location.path.clone(),
-                                patch: suggestion.patch.clone(),
-                                original_code,
-                                modified_code,
-                            });
-                        } else {
-                            warn!("Failed to apply patch: {}", suggestion.patch);
-                        }
-                    } else {
-                        warn!("Failed to parse patch: {}", suggestion.patch);
-                    }
-                }
-            }
-        }
-
-        if patch_candidates.is_empty() {
-            return Ok(());
-        }
-
-        writeln!(writer)?;
-        writeln!(
-            writer,
-            "{}{}{}",
-            style(" AUTOFIXES: ").bold().reverse(),
-            style(patch_candidates.len().to_formatted_string(&Locale::en))
-                .bold()
-                .reverse(),
-            style(" ").bold().reverse()
-        )?;
-        writeln!(writer)?;
-
-        for candidate in patch_candidates {
-            let diff = TextDiff::from_lines(&candidate.original_code, &candidate.modified_code);
-            let mut patch_writer = vec![];
-
-            for (idx, group) in diff.grouped_ops(3).iter().enumerate() {
-                if idx > 0 {
-                    writeln!(patch_writer, "{:-^1$}", "-", 80)?;
-                }
-                for op in group {
-                    for change in diff.iter_inline_changes(op) {
-                        let (sign, s) = match change.tag() {
-                            ChangeTag::Delete => ("-", Style::new().red()),
-                            ChangeTag::Insert => ("+", Style::new().green()),
-                            ChangeTag::Equal => (" ", Style::new().dim()),
-                        };
-                        write!(
-                            patch_writer,
-                            "{}{} |{}",
-                            style(Line(change.old_index())).dim(),
-                            style(Line(change.new_index())).dim(),
-                            s.apply_to(sign).bold(),
-                        )?;
-                        for (emphasized, value) in change.iter_strings_lossy() {
-                            if emphasized {
-                                write!(
-                                    patch_writer,
-                                    "{}",
-                                    s.apply_to(value).underlined().on_black()
-                                )?;
-                            } else {
-                                write!(patch_writer, "{}", s.apply_to(value))?;
-                            }
-                        }
-                        if change.missing_newline() {
-                            writeln!(patch_writer)?;
-                        }
-                    }
-                }
-            }
-
-            // For a reason that I haven't figured out yet, sometimes we print
-            // empty patches. This is a workaround to skip those issues.
-            if !patch_writer.is_empty() {
-                let start_line = candidate.issue.range().unwrap_or_default().start_line;
-
-                writeln!(
-                    writer,
-                    "{}{}",
-                    style(&candidate.path).underlined(),
-                    style(format!(":{}", start_line)).dim()
-                )?;
-
-                writeln!(
-                    writer,
-                    "{} {}",
-                    formatted_level(candidate.issue.level()),
-                    style(candidate.issue.message.replace('\n', " ").trim())
-                )?;
-
-                write!(writer, "{}", String::from_utf8_lossy(&patch_writer))?;
-                writeln!(
-                    writer,
-                    "{} {}",
-                    formatted_source(&candidate.issue),
-                    match candidate.source {
-                        SuggestionSource::Llm => format!("[{}]", style("ai fix").cyan()),
-                        _ => "".to_string(),
-                    }
-                )?;
-                writeln!(writer)?;
-
-                if std::io::stdin().is_terminal() {
-                    match self.apply_mode {
-                        ApplyMode::None => {} // Skip and don't ask
-                        ApplyMode::All => {
-                            apply_fix(writer, &candidate)?;
-                        }
-                        ApplyMode::Ask => {
-                            let mut answered = false;
-
-                            // Loop until we get a valid answer
-                            while !answered {
-                                if let Ok(answer) = prompt_apply_this_fix() {
-                                    match answer.as_str() {
-                                        "Y" | "y" | "yes" => {
-                                            answered = true;
-                                            apply_fix(writer, &candidate)?;
-                                        }
-                                        "A" | "a" | "all" => {
-                                            answered = true;
-                                            self.apply_mode = ApplyMode::All;
-                                            apply_fix(writer, &candidate)?;
-                                        }
-                                        "N" | "n" | "no" => {
-                                            answered = true;
-                                        }
-                                        "none" => {
-                                            answered = true;
-                                            self.apply_mode = ApplyMode::None;
-                                        }
-                                        _ => {}
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                writeln!(writer)?;
-            }
-        }
-
-        Ok(())
-    }
-
     pub fn print_conclusion(&self, writer: &mut dyn std::io::Write) -> Result<()> {
         if self.verbose >= 1 && self.report.targets_count() > 0 {
             self.print_processed_files(writer)?;
@@ -480,66 +298,6 @@ impl TextFormatter {
         )?;
 
         Ok(())
-    }
-}
-
-fn prompt_apply_this_fix() -> Result<String> {
-    Ok(Input::<String>::with_theme(&ColorfulTheme::default())
-        .with_prompt("Apply this fix? [Yes/no/all/none]")
-        .default("Y".to_string())
-        .show_default(false)
-        .allow_empty(true)
-        .interact_text()?)
-}
-
-fn apply_fix(writer: &mut dyn std::io::Write, candidate: &PatchCandidate) -> Result<()> {
-    if let Ok(patch) = Patch::from_str(&candidate.patch) {
-        if let Ok(modified_code) = diffy::apply(&candidate.original_code, &patch) {
-            std::fs::write(&candidate.path, &modified_code)
-                .with_context(|| format!("Failed to apply path to file: {}", candidate.path))?;
-
-            eprintln!(
-                "{} {}",
-                style("✔ Fixed:").green().bold(),
-                style(&candidate.path).underlined()
-            );
-        } else {
-            warn!("Failed to apply patch: {}", candidate.patch);
-            writeln!(
-                writer,
-                "{} {}",
-                style("Failed to apply patch:").red(),
-                style(&candidate.path).underlined()
-            )?;
-        }
-    } else {
-        warn!("Failed to parse patch: {}", candidate.patch);
-        writeln!(
-            writer,
-            "{} {}",
-            style("Failed to parse patch:").red(),
-            style(&candidate.path).underlined()
-        )?;
-    }
-
-    Ok(())
-}
-
-fn formatted_level(level: Level) -> String {
-    match level {
-        Level::High => style("high  ").red().to_string(),
-        Level::Medium => style("medium").magenta().to_string(),
-        Level::Low => style("low   ").yellow().to_string(),
-        Level::Fmt => style("fmt   ").dim().to_string(),
-        _ => format!("{:?}", level),
-    }
-}
-
-fn formatted_source(issue: &Issue) -> String {
-    if !issue.rule_key.is_empty() {
-        format!("{}", style(issue.rule_id()).dim())
-    } else {
-        format!("{}", style(issue.tool.clone()).dim())
     }
 }
